requirements.txt
dify-plugin==0.6.2
requests>=2.31.0
sseclient-py>=1.8.0
